Create reusable Projects component
